requirements.txt
# UI
nicegui
fastapi
uvicorn

# AI
python-dotenv
openai
openai[voice_helpers]
numpy

# Utils
mypy
pylint
black
pytest
pytest-cov